Change type of showLensFlare variable
It was a double, but should be an integer

+++ src/simulation/vizard/_GeneralModuleFiles/vizStructures.h
@@ -216,7 +216,7 @@ Light
     double markerDiameter=-1;                       //!< [m] (Optional) Size to draw the visible lens of the light, default is 0.01 m
     std::vector<int> color;                         //!< (Optional) Send desired RGBA as values between 0 and 255, default is pure white
     double gammaSaturation=-1;                      //!< (Optional) Desired gamma saturation of the light lens, 0 to match light color, 1.0 for pure white, default is 0.8
-    double showLensFlare;                           //!< (Optional) Value of 0 (protobuffer default) to use viz default, -1 for false, 1 for true
+    int showLensFlare;                              //!< (Optional) Value of 0 (protobuffer default) to use viz default, -1 for false, 1 for true
     double lensFlareBrightness=-1;                  //!< (Optional) Simulates refraction of light in camera lens, this value controls the size and brightness of the lens flare, default is 0.3
     double lensFlareFadeSpeed=-1;                   //!< (Optional) Speed with which the lens flare fades, default is 4.0
     ReadFunctor<DeviceCmdMsgPayload> onOffCmdInMsg; //!< (Optional)  incoming light on/off cmd state msg
